Apply Dart library pattern.
