+++ lib/src/upload_method.dart
@@ -1,0 +1,7 @@
+part of flutter_uploader;
+
+enum UploadMethod {
+  POST,
+  PUT,
+  PATCH,
+}
